Use PositiveIntegerField for schema_id fields

+++ netbox_custom_objects/migrations/0005_customobjecttype_next_schema_id_and_more.py
@@ -15,7 +15,7 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='customobjecttype',
             name='next_schema_id',
-            field=models.PositiveSmallIntegerField(default=0, editable=False),
+            field=models.PositiveIntegerField(default=0, editable=False),
         ),
         migrations.AddField(
             model_name='customobjecttype',
@@ -40,7 +40,7 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='customobjecttypefield',
             name='schema_id',
-            field=models.PositiveSmallIntegerField(blank=True, null=True),
+            field=models.PositiveIntegerField(blank=True, null=True),
         ),
         migrations.AlterField(
             model_name='customobjecttype',
